backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
langchain==0.1.9
langchain-openai==0.0.8
openai==1.12.0
serpapi==0.1.5
beautifulsoup4==4.12.3
requests==2.31.0
pydantic>=2.6.3
pydantic-core>=2.16.3
httpx==0.27.2 